+++ app/src/main/java/com/main/prodapp/fragments/calendar/CalendarAdapter.java
@@ -1,4 +1,4 @@
-package com.main.prodapp;
+package com.main.prodapp.fragments.calendar;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -6,6 +6,8 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.main.prodapp.R;
 
 import java.util.ArrayList;
