+++ app/src/main/java/com/main/prodapp/fragments/calendar/CalendarListAdapter.kt
@@ -1,0 +1,4 @@
+package com.main.prodapp.fragments.calendar
+
+class CalendarListAdapter {
+}
