+++ app/src/main/java/com/main/prodapp/fragments/calendar/CalendarViewHolder.java
@@ -1,10 +1,12 @@
-package com.main.prodapp;
+package com.main.prodapp.fragments.calendar;
 
 import android.view.View;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.main.prodapp.R;
 
 public class CalendarViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
     public TextView dayOfMonth;
